requirements.txt
requests
pyyaml
pandas
openai
